task 6: update

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -5,10 +5,11 @@ a type-annotated function sum_mixed_list
 which takes a list mxd_lst of integers and floats
 and returns their sum as a float.
 """
-from typing import List, Union
+from typing import List
+from typing import Union
 
 
-def sum_mixed_list(mxd_lst: List[Union[float, int]]) -> float:
+def sum_mixed_list(mxd_lst: List[Union[int, float]]) -> float:
     """
     adds ints and floats in a list
     returns the sum, return type: float
